rename some routes
