+++ src/app/routes/index.js
@@ -1,8 +1,8 @@
 import article from './article';
 import cpsAsset from './cpsAsset';
-import error200 from './200';
-import error404 from './404';
-import index from './index/index';
+import error200 from './error200';
+import error404 from './error404';
+import home from './home';
 import radio from './radio';
 
-export default [article, cpsAsset, error200, error404, index, radio];
+export default [article, cpsAsset, error200, error404, home, radio];
